refactor: update module imports and restructure filter and utility exports
