--- packages/core/src/filters/index.ts
@@ -1,2 +1,0 @@
-export * from "./biquadratic-filter";
-export * from "./finite-impulse-response-filter";
