--- packages/core/src/processors/index.ts
@@ -1,1 +1,0 @@
-export * from "./loudness-processor";
